+++ app/js/client.js
@@ -5,7 +5,7 @@ require('angular/angular');
 var mmmApp = angular.module('mmmApp', []);
 
 //services
-require('./services/resource_backend_service')(mmmApp);
+require('./services/mmm_service')(mmmApp);
 
 //controllers
 require('./controllers/mmm_controller')(mmmApp);
