+++ app/js/controllers/mmm_controller.js
@@ -2,18 +2,14 @@
 /*thanks to S. Burbidge, J. Elsey, S. Lingwood, & C. Renwick for help*/
 
 module.exports = function(app) {
-  app.controller('mmmCtrl', ['$scope', '$http', 'ResourceBackend',
-    function($scope, $http, ResourceBackend) {
-
-      var mmmBackend = new ResourceBackend();
+  app.controller('mmmCtrl', ['$scope', 'mmmBuilder',
+    function($scope, mmmBuilder) {
 
       $scope.calcMMM = function() {
-        mmmBackend.postMMM ($scope.numInput.split(' '))
-        .success(function(data) {
-          $scope.mean = data.mean;
-          $scope.median = data.median;
-          $scope.mode = data.mode;
-        });
+        var data = mmmBuilder.buildMMM($scope.numInput.split(' '));
+        $scope.mean = data.mean;
+        $scope.median = data.median;
+        $scope.mode = data.mode;
       };
     }]);
 };
