+++ test/server/mean_median_mode_tests.js
@@ -1,23 +1,23 @@
-'use strict';
+// 'use strict';
 
-var chai = require('chai');
-var expect = chai.expect;
-var chaihttp = require('chai-http');
+// var chai = require('chai');
+// var expect = chai.expect;
+// var chaihttp = require('chai-http');
 
-chai.use(chaihttp);
-require('../../server');
+// chai.use(chaihttp);
+// require('../../server');
 
-describe('Mean Median & Mode', function() {
-  it('should get accurate mean, median, and mode values', function(done) {
-    chai.request('http://localhost:3000')
-    .post('/api/mmm')
-    .send({numList: [3, 5, 7, 9]})
-    .end(function(err, res) {
-      expect(err).to.eql(null);
-      expect(res.body.mean).to.equal(6);
-      expect(res.body.median).to.equal(6);
-      expect(res.body.mode).to.equal(3);
-      done();
-    });
-  });
-});
+// describe('Mean Median & Mode', function() {
+//   it('should get accurate mean, median, and mode values', function(done) {
+//     chai.request('http://localhost:3000')
+//     .post('/api/mmm')
+//     .send({numList: [3, 5, 7, 9]})
+//     .end(function(err, res) {
+//       expect(err).to.eql(null);
+//       expect(res.body.mean).to.equal(6);
+//       expect(res.body.median).to.equal(6);
+//       expect(res.body.mode).to.equal(3);
+//       done();
+//     });
+//   });
+// });
